Move VCR test to test_utils package
