+++ src/command_modules/azure-cli-keyvault/azure/cli/command_modules/keyvault/tests/test_keyvault_commands.py
@@ -8,8 +8,8 @@
 #pylint: disable=bad-continuation
 import os
 
-from azure.cli.core.utils.vcr_test_base import (ResourceGroupVCRTestBase, JMESPathCheck,
-                                           NoneCheck)
+from azure.cli.core.test_utils.vcr_test_base import (ResourceGroupVCRTestBase, JMESPathCheck,
+                                                     NoneCheck)
 
 TEST_DIR = os.path.abspath(os.path.join(os.path.abspath(__file__), '..'))
