+++ src/command_modules/azure-cli-resource/azure/cli/command_modules/resource/tests/test_resource.py
@@ -7,8 +7,10 @@ import os
 import time
 # AZURE CLI RESOURCE TEST DEFINITIONS
 
-from azure.cli.core.utils.vcr_test_base import (VCRTestBase, JMESPathCheck, NoneCheck, BooleanCheck,
-                                                ResourceGroupVCRTestBase, MOCKED_SUBSCRIPTION_ID)
+from azure.cli.core.test_utils.vcr_test_base import (VCRTestBase, JMESPathCheck, NoneCheck,
+                                                     BooleanCheck,
+                                                     ResourceGroupVCRTestBase,
+                                                     MOCKED_SUBSCRIPTION_ID)
 
 #pylint: disable=method-hidden, line-too-long
 class ResourceGroupScenarioTest(VCRTestBase): # Not RG test base because it tests the actual deletion of a resource group
