+++ src/command_modules/azure-cli-role/azure/cli/command_modules/role/tests/test_graph.py
@@ -3,7 +3,7 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 #---------------------------------------------------------------------------------------------
 
-from azure.cli.core.utils.vcr_test_base import VCRTestBase, JMESPathCheck, NoneCheck
+from azure.cli.core.test_utils.vcr_test_base import VCRTestBase, JMESPathCheck, NoneCheck
 
 class ServicePrincipalExpressCreateScenarioTest(VCRTestBase):
