+++ src/command_modules/azure-cli-role/azure/cli/command_modules/role/tests/test_role.py
@@ -10,7 +10,7 @@ import re
 import tempfile
 import time
 
-from azure.cli.core.utils.vcr_test_base import VCRTestBase, JMESPathCheck, ResourceGroupVCRTestBase, NoneCheck, MOCKED_SUBSCRIPTION_ID
+from azure.cli.core.test_utils.vcr_test_base import VCRTestBase, JMESPathCheck, ResourceGroupVCRTestBase, NoneCheck, MOCKED_SUBSCRIPTION_ID
 
 class RoleCreateScenarioTest(VCRTestBase):
