+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/tests/test_storage.py
@@ -14,7 +14,7 @@ import time
 
 from six import StringIO
 
-from azure.cli.core.utils.vcr_test_base import \
+from azure.cli.core.test_utils.vcr_test_base import \
     (VCRTestBase, ResourceGroupVCRTestBase, StorageAccountVCRTestBase, 
      JMESPathCheck, NoneCheck, BooleanCheck, StringCheck)
 from azure.cli.core._util import CLIError
