+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/tests/test_vm_commands.py
@@ -9,10 +9,10 @@ import os
 import tempfile
 import platform
 
-from azure.cli.core.utils.vcr_test_base import (VCRTestBase,
-                                                ResourceGroupVCRTestBase,
-                                                JMESPathCheck,
-                                                NoneCheck)
+from azure.cli.core.test_utils.vcr_test_base import (VCRTestBase,
+                                                     ResourceGroupVCRTestBase,
+                                                     JMESPathCheck,
+                                                     NoneCheck)
 
 TEST_DIR = os.path.abspath(os.path.join(os.path.abspath(__file__), '..'))
